model_training/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python
numpy
pillow
matplotlib
scikit-image
pyyaml
tqdm
pytest
